+++ routes/error.jsx
@@ -1,9 +1,6 @@
-// modules
-import React from 'react';
-import { renderToString } from 'react-dom/server';
 // react components
+import reaxpressResponseHandler from './reaxpressResponseHandler';
 import Page from '../src/react/App/_global/Page';
-import template from '../template';
 
 const router = require('express').Router();
 
@@ -15,15 +12,13 @@ router.use((req, res, next) => {
 // error handler
 router.use((req, res) => {
   const status = res.locals.status || 500;
-  res.status(status);
-  const rd = res.locals.reaxpressData;
-  rd.page = {
+  const reaxpressData = res.locals.reaxpressData;
+  reaxpressData.status = status;
+  reaxpressData.page = {
     title: `Error: ${status}`,
     content: 'There was a problem processing your request or this page simply does not exist.',
   };
-  const isReaxpress = req.originalUrl.indexOf('reaxpress=true') > -1;
-  if (isReaxpress) { return res.json(rd); }
-  return res.send(template(rd, renderToString(<Page reaxpressData={rd} />)));
+  reaxpressResponseHandler(req, res, Page, reaxpressData, status);
 });
 
 module.exports = router;